彻底删除 numeric 数据类型并重写所有耦合代码
- 删除 numeric 类型定义、numeric.c 实现及 numeric.h 头文件
- 删除 numeric 相关 catalog 条目(pg_type/pg_proc/pg_cast/pg_opfamily 等)
- 删除 numeric 回归测试(numeric.sql/numeric_big.sql 及期望)
- 重写所有依赖 numeric 的代码为 float8/int8:
  date.c/timestamp.c/formatting.c/dbsize.c/pg_lsn.c/
  format_type.c/pgstatfuncs.c/ruleutils.c/selfuncs.c/
  rangetypes.c/fe_utils/print.c/parse_node.c/xlogfuncs.c
- 重写整数聚合为 _int8 内部状态表示(intagg.c),注册 int8_avg 及 avg(int2/4/8)
- EXTRACT/date_part/epoch 改为返回 float8,pg_lsn +/- 与 pg_size_pretty 改为 float8/int64
- 更新 23 个回归 + 7 个隔离测试用例期望文件
- 同步更新 mydoc/CHANGE.md

make check-world 全量回归测试通过(回归 82/82, 隔离 66/66)

+++ src/test/regress/sql/arrays.sql
@@ -210,7 +210,7 @@ select * from arrtest1;
 --
 
 -- table creation and INSERTs
-CREATE TABLE arrtest2 (i integer ARRAY[4], f float8[], n numeric[], t text[], d timestamp[]);
+CREATE TABLE arrtest2 (i integer ARRAY[4], f float8[], n int8[], t text[], d timestamp[]);
 INSERT INTO arrtest2 VALUES(
   ARRAY[[[113,142],[1,147]]],
   ARRAY[1.1,1.2,1.3]::float8[],
@@ -627,7 +627,7 @@ select array_agg(ar)
 select unnest(array[1,2,3]);
 select * from unnest(array[1,2,3]);
 select unnest(array[1,2,3,4.5]::float8[]);
-select unnest(array[1,2,3,4.5]::numeric[]);
+select unnest(array[1,2,3,4.5]::int8[]);
 select unnest(array[1,2,3,null,4,null,null,5,6]);
 select unnest(array[1,2,3,null,4,null,null,5,6]::text[]);
 select abs(unnest(array[1,2,null,-3]));
@@ -701,9 +701,9 @@ drop type textandtext;
 -- this exercises the varwidth and float8 code paths
 SELECT
     op,
-    width_bucket(op::numeric, ARRAY[1, 3, 5, 10.0]::numeric[]) AS wb_n1,
-    width_bucket(op::numeric, ARRAY[0, 5.5, 9.99]::numeric[]) AS wb_n2,
-    width_bucket(op::numeric, ARRAY[-6, -5, 2.0]::numeric[]) AS wb_n3,
+    width_bucket(op::int8, ARRAY[1, 3, 5, 10.0]::int8[]) AS wb_n1,
+    width_bucket(op::int8, ARRAY[0, 5.5, 9.99]::int8[]) AS wb_n2,
+    width_bucket(op::int8, ARRAY[-6, -5, 2.0]::int8[]) AS wb_n3,
     width_bucket(op::float8, ARRAY[1, 3, 5, 10.0]::float8[]) AS wb_f1,
     width_bucket(op::float8, ARRAY[0, 5.5, 9.99]::float8[]) AS wb_f2,
     width_bucket(op::float8, ARRAY[-6, -5, 2.0]::float8[]) AS wb_f3
