+++ src/test/regress/sql/float4.sql
@@ -52,7 +52,7 @@ SELECT 'Infinity'::float4 / 'Infinity'::float4;
 SELECT '42'::float4 / 'Infinity'::float4;
 SELECT 'nan'::float4 / 'nan'::float4;
 SELECT 'nan'::float4 / '0'::float4;
-SELECT 'nan'::numeric::float4;
+SELECT 'nan'::int8::float4;
 
 SELECT * FROM FLOAT4_TBL;
